Move spack runner closer to spack package manager
