+++ pytest.ini
@@ -1,7 +1,7 @@
 # content of pytest.ini
 [pytest]
 addopts = --durations=30 -ra
-testpaths = lib/ramble/ramble/test
+testpaths = lib/ramble/ramble/test, var/ramble/repos/*/*/*/test
 python_files = *.py
 filterwarnings =
   ignore::UserWarning
